Integrate new type.Numeric<*> spec through rest of API

+++ maven/plugin/xdb-maven-plugin/src/test/java/org/safris/maven/plugin/xdb/xde/SetFunctionTest.java
@@ -22,6 +22,7 @@ import static org.safris.xdb.entities.DML.MIN;
 import static org.safris.xdb.entities.DML.SELECT;
 
 import java.io.IOException;
+import java.math.BigInteger;
 import java.sql.SQLException;
 
 import org.junit.Assert;
@@ -48,9 +49,9 @@ public class SetFunctionTest extends LoggableTest {
       execute();
 
     Assert.assertTrue(rows.nextRow());
-    Assert.assertEquals(Long.valueOf(24367857008l), rows.nextEntity().get());
+    Assert.assertEquals(BigInteger.valueOf(24367857008l), rows.nextEntity().get());
     Assert.assertEquals("White Plains", rows.nextEntity().get());
     Assert.assertEquals(classicmodels.Address.Country.AU, rows.nextEntity().get());
-    Assert.assertEquals(Long.valueOf(21003), rows.nextEntity().get());
+    Assert.assertEquals(BigInteger.valueOf(21003), rows.nextEntity().get());
   }
 }
